backend/requirements.txt
# ─────────────────────────────────────────
# Web Framework
# ─────────────────────────────────────────
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.11.7
python-multipart==0.0.9        # required for FastAPI file uploads (UploadFile)

# ─────────────────────────────────────────
# Auth
# ─────────────────────────────────────────
python-jose[cryptography]==3.3.0   # JWT token generation & verification
bcrypt==4.0.1                     # password hashing
passlib[bcrypt]==1.7.4             # bcrypt wrapper used with FastAPI

# ─────────────────────────────────────────
# Database
# ─────────────────────────────────────────
supabase==2.24.0

# ─────────────────────────────────────────
# HTTP Client (weather API)
# ─────────────────────────────────────────
httpx==0.27.0

# ─────────────────────────────────────────
# ML — Core
# ─────────────────────────────────────────
scikit-learn==1.5.2
xgboost==2.1.1
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2

# ─────────────────────────────────────────
# ML — Disease Detection
# ─────────────────────────────────────────
torch==2.3.0
torchvision==0.18.0
Pillow==10.3.0
huggingface_hub==0.23.0

# ─────────────────────────────────────────
# Utils
# ─────────────────────────────────────────
python-dotenv==1.0.1


twilio==8.5.0